+++ lib/openstudio/workflow/adapters/local.rb
@@ -76,10 +76,10 @@ module OpenStudio
         end
 
         def communicate_results(directory, results)
-          zip_results(directory, 'workflow')
+          zip_results(directory)
 
           if results.is_a? Hash
-            File.open("#{directory}/datapoint_out.json", 'w') { |f| f << JSON.pretty_generate(results) }
+            File.open("#{directory}/data_point_out.json", 'w') { |f| f << JSON.pretty_generate(results) }
           else
             pp "Unknown datapoint result type. Please handle #{results.class}"
             # data_point_json_path = OpenStudio::Path.new(run_dir) / OpenStudio::Path.new('data_point_out.json')
@@ -90,32 +90,10 @@ module OpenStudio
 
         # For the local adapter send back a handle to a file to append the data. For this adapter
         # the log messages are likely to be the same as the run.log messages.
-        # TODO: do we really want two local logs from the Local adapter? One is in the run dir and the other is in the root
+        # ?: do we really want two local logs from the Local adapter? One is in the run dir and the other is in the root
         def get_logger(directory, _options = {})
           @log ||= File.open("#{directory}/local_adapter.log", 'w')
           @log
-        end
-
-        # TODO: this uses a system call to zip results at the moment
-        def zip_results(directory, _analysis_type = 'workflow')
-          current_dir = Dir.pwd
-          begin
-            # create zip file using a system call
-            # @logger.info "Zipping up data point #{analysis_dir}"
-            if File.directory? directory
-              Dir.chdir(directory)
-              `zip -9 -r --exclude=*.rb* data_point.zip .`
-            end
-
-            # zip up only the reports folder
-            report_dir = 'reports'
-            # @logger.info "Zipping up Analysis Reports Directory #{report_dir}/reports"
-            if File.directory? report_dir
-              `zip -9 -r data_point_reports.zip reports`
-            end
-          ensure
-            Dir.chdir(current_dir)
-          end
         end
       end
     end